convert timestamp to double value

+++ app/src/main/java/com/ping/android/activity/AddGroupActivity.java
@@ -270,7 +270,7 @@ public class AddGroupActivity extends CoreActivity implements View.OnClickListen
     }
 
     private void createGroup(List<User> toUsers, String groupKey, String msg, String profileImage) {
-        long timestamp = System.currentTimeMillis() / 1000L;
+        double timestamp = System.currentTimeMillis() / 1000L;
         Group group = new Group();
         group.timestamp = timestamp;
         group.groupName = etGroupName.getText().toString().trim();
@@ -299,7 +299,7 @@ public class AddGroupActivity extends CoreActivity implements View.OnClickListen
     private void openPicker() {
         String profileFileFolder = getExternalFilesDir(null).getAbsolutePath() + File.separator +
                 "profile" + File.separator + fromUser.key;
-        long timestamp = System.currentTimeMillis() / 1000L;
+        double timestamp = System.currentTimeMillis() / 1000L;
         String profileFileName = "" + timestamp + "-" + fromUser.key + ".png";
         String profileFilePath = profileFileFolder + File.separator + profileFileName;
         imagePickerHelper = ImagePickerHelper.from(this)
